vector-db/requirements.txt
numpy
matplotlib
requests
pyjson
weaviate-client
python-dotenv